another commit (done many changes, don't know how to name it)

+++ src/include/connect.h
@@ -1,4 +1,11 @@
-int	InitConnection(int *, char *, char *);
-int	login(int, char *);
-int	join(int, char *);
-int	pong(int, char*);
+#include <netdb.h>
+#include <sys/socket.h>
+
+#include <stdio.h>
+#include <string.h>
+#include <unistd.h>
+
+int InitConnection(int *, char *, char *);
+ssize_t login(int, char *);
+ssize_t join(int, char *);
+ssize_t pong(int, char *);
